DSS-1677: Review Fix 1. Integer to Long, service modification for same

+++ api/cp-dss-api/gateway/src/main/java/com/cpdss/gateway/domain/Voyage.java
@@ -56,5 +56,5 @@ public class Voyage {
 
   private Long statusId;
 
-  private Integer timezoneId;
+  private Long timezoneId;
 }
