+++ api/cp-dss-api/gateway/src/main/java/com/cpdss/gateway/service/LoadableStudyService.java
@@ -246,7 +246,7 @@ public class LoadableStudyService {
             .setVoyageNo(voyage.getVoyageNo())
             .setStartDate(!StringUtils.isEmpty(voyage.getStartDate()) ? voyage.getStartDate() : "")
             .setEndDate(!StringUtils.isEmpty(voyage.getEndDate()) ? voyage.getEndDate() : "")
-            .setTimezoneId(voyage.getTimezoneId() != null ? voyage.getTimezoneId() : 0)
+            .setTimezoneId(voyage.getTimezoneId() != null ? voyage.getTimezoneId().intValue() : 0)
             .build();
 
     VoyageReply voyageReply = this.saveVoyage(voyageRequest);
